Refactor playback buttons to a separate component
[#164420580]
