+++ src/components/bottom-bar.scss
@@ -47,28 +47,4 @@ $overflowHeight: 14px;
     margin-left: 5px;
     margin-right: 5px;
   }
-
-  .tallButton {
-    min-width: 60px;
-    height: 100%;
-    border-radius: 0;
-    &:first-child {
-      border-top-left-radius: 10px;
-    }
-    &:last-child {
-      border-top-right-radius: 10px;
-    }
-    &:hover {
-      background: #f2f2f2;
-    }
-    svg {
-      display: block;
-      height: 30px;
-      width: 30px;
-      margin: 0 auto;
-      path {
-        fill: #666666;
-      }
-    }
-  }
 }
