+++ src/components/bottom-bar.test.tsx
@@ -4,7 +4,7 @@ import { createStores } from "../models/stores";
 import { Provider } from "mobx-react";
 import { BottomBar } from "./bottom-bar";
 import { SeasonButton } from "./season-button";
-import Button from "@material-ui/core/Button";
+import { PlaybackButtons } from "./playback-buttons";
 
 describe("BottomBar component", () => {
   let stores = createStores();
@@ -18,32 +18,7 @@ describe("BottomBar component", () => {
         <BottomBar />
       </Provider>
     );
-    expect(wrapper.find(Button).length).toEqual(3);
+    expect(wrapper.find(PlaybackButtons).length).toEqual(1);
     expect(wrapper.find(SeasonButton).length).toEqual(1);
-  });
-
-  it("start button is disabled until model is ready", () => {
-    const wrapper = mount(
-      <Provider stores={stores}>
-        <BottomBar />
-      </Provider>
-    );
-    const start = wrapper.find('[data-test="start-button"]').first();
-    expect(start.prop("disabled")).toEqual(true);
-  });
-
-  describe("restart button", () => {
-    it("resets both simulation and the view", () => {
-      jest.spyOn(stores.simulation, "reset");
-      jest.spyOn(stores.ui, "setNorthAtlanticView");
-      const wrapper = mount(
-        <Provider stores={stores}>
-          <BottomBar />
-        </Provider>
-      );
-      wrapper.find('[data-test="restart-button"]').first().simulate("click");
-      expect(stores.simulation.reset).toHaveBeenCalled();
-      expect(stores.ui.setNorthAtlanticView).toHaveBeenCalled();
-    });
   });
 });
