+++ src/components/bottom-bar.tsx
@@ -2,10 +2,8 @@ import { inject, observer } from "mobx-react";
 import * as React from "react";
 import { BaseComponent, IBaseProps } from "./base";
 import Button from "@material-ui/core/Button";
-import StartIcon from "../assets/start.svg";
-import PauseIcon from "../assets/pause.svg";
-import RestartIcon from "../assets/restart.svg";
 import { SeasonButton } from "./season-button";
+import { PlaybackButtons } from "./playback-buttons";
 
 import * as css from "./bottom-bar.scss";
 
@@ -17,7 +15,6 @@ interface IState {}
 export class BottomBar extends BaseComponent<IProps, IState> {
 
   public render() {
-    const sim = this.stores.simulation;
     const ui = this.stores.ui;
     return (
       <div className={css.bottomBar}>
@@ -25,17 +22,7 @@ export class BottomBar extends BaseComponent<IProps, IState> {
           <SeasonButton />
         </div>
         <div className={css.widgetGroup}>
-          <Button
-            onClick={sim.simulationStarted ? sim.stop : sim.start}
-            disabled={!sim.ready}
-            className={css.tallButton}
-            data-test="start-button"
-          >
-            { sim.simulationStarted ? <span><PauseIcon/> Stop</span> : <span><StartIcon /> Start</span> }
-          </Button>
-          <Button className={css.tallButton} data-test="restart-button" onClick={this.handleReset}>
-            <span><RestartIcon/> Restart</span>
-          </Button>
+          <PlaybackButtons />
         </div>
         <div className={css.seasonSelect}>
         {
@@ -45,10 +32,5 @@ export class BottomBar extends BaseComponent<IProps, IState> {
         </div>
       </div>
     );
-  }
-
-  public handleReset = () => {
-    this.stores.simulation.reset();
-    this.stores.ui.setNorthAtlanticView();
   }
 }
