+++ src/components/playback-buttons.scss
@@ -1,0 +1,23 @@
+.playbackButton {
+  min-width: 60px;
+  height: 100%;
+  border-radius: 0;
+  &:first-child {
+    border-top-left-radius: 10px;
+  }
+  &:last-child {
+    border-top-right-radius: 10px;
+  }
+  &:hover {
+    background: #f2f2f2;
+  }
+  svg {
+    display: block;
+    height: 30px;
+    width: 30px;
+    margin: 0 auto;
+    path {
+      fill: #666666;
+    }
+  }
+}
